converting to jdo

+++ trunk/Core/src/com/gawkat/core/server/db/oauth/ApplicationData.java
@@ -6,7 +6,7 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class ApplicationData implements IsSerializable {
 
-	public int applicationId;
+	public Long applicationId;
 	public String consumerKey;
 	public String consumerSecret;
